backend/requirements.txt
fastapi==0.95.2
uvicorn==0.15.0
python-multipart==0.0.5
ollama==0.1.0
Pillow==9.3.0
python-docx==0.8.11
PyPDF2==3.0.1
pydantic==1.10.7
